Proper fix for bug #1920869 "HoF: palette artifacts in inventory".
svn-id: r31208

+++ engines/kyra/screen_v2.h
@@ -54,7 +54,7 @@ public:
 	void copyWsaRect(int x, int y, int w, int h, int dimState, int plotFunc, const uint8 *src,
 					int unk1, const uint8 *unkPtr1, const uint8 *unkPtr2);
 
-	uint8 *generateOverlay(const uint8 *palette, uint8 *buffer, int color, int factor);
+	uint8 *generateOverlay(const uint8 *palette, uint8 *buffer, int color, uint16 factor);
 	void applyOverlay(int x, int y, int w, int h, int pageNum, const uint8 *overlay);
 
 	// shape handling
